refactor(client): extract test_helpers to sibling file
Move pub(super) mod test_helpers { ... } from src/client/mod.rs
into its own src/client/test_helpers.rs file, declared as
`#[cfg(test)] pub(super) mod test_helpers;`. No #[path] directive
needed because the filename matches the module name. Visibility
(pub(super) and pub fn) is preserved across the file boundary;
existing imports in src/client/*_tests.rs remain valid.

Wave 1 of the test-sibling migration. See
docs/superpowers/specs/2026-05-05-test-sibling-migration-design.md.
